Update for ES cluster of 3 master and 2 data nodes
Creates an Elasticsearch cluster named "monitoring" in default namespace, consisting of 3 master nodes and 2 data nodes. The service is exposed as an internal LoadBalancer type and all the traffic for ES is terminated using "mycompany-tls" certificate. Additional configurations are set for OCI such as oci-bv, annotations, etc. Node selectors are optional, if using, make sure the labels are available on host nodes.